# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Server for SAC App",
        "title": "SAC API",
        "contact": {
            "name": "David Oduneye and Garrett Ladley",
            "email": "generatesac@gmail.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/category/": {
            "get": {
                "description": "Retrieves all categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Retrieve all categories",
                "operationId": "get-categories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Creates a category",
                "operationId": "create-category",
                "parameters": [
                    {
                        "description": "Category Body",
                        "name": "categoryBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CategoryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/category/{categoryID}": {
            "get": {
                "description": "Retrieves a category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Retrieve a category",
                "operationId": "get-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Updates a category",
                "operationId": "update-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category Body",
                        "name": "categoryBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CategoryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Deletes a category",
                "operationId": "delete-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/club/{clubID}/members": {
            "get": {
                "description": "Retrieves all members associated with a club",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "club-member"
                ],
                "summary": "Retrieve all members for a club",
                "operationId": "get-members-by-club",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club ID",
                        "name": "clubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.CategoryRequestBody": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "models.Club": {
            "type": "object",
            "required": [
                "application_link",
                "description",
                "is_recruiting",
                "name",
                "num_members",
                "preview",
                "recruitment_cycle",
                "recruitment_type"
            ],
            "properties": {
                "application_link": {
                    "type": "string",
                    "maxLength": 255
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "description": {
                    "description": "MongoDB URL",
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "is_recruiting": {
                    "type": "boolean"
                },
                "logo": {
                    "description": "S3 URL",
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "num_members": {
                    "type": "integer",
                    "minimum": 1
                },
                "preview": {
                    "type": "string",
                    "maxLength": 255
                },
                "recruitment_cycle": {
                    "maxLength": 255,
                    "enum": [
                        "fall",
                        "spring",
                        "fallSpring",
                        "always"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RecruitmentCycle"
                        }
                    ]
                },
                "recruitment_type": {
                    "maxLength": 255,
                    "enum": [
                        "unrestricted",
                        "tryout",
                        "application"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RecruitmentType"
                        }
                    ]
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.College": {
            "type": "string",
            "enum": [
                "CAMD",
                "DMSB",
                "KCCS",
                "CE",
                "BCHS",
                "SL",
                "CPS",
                "CS",
                "CSSH"
            ],
            "x-enum-comments": {
                "BCHS": "Bouvé College of Health Sciences",
                "CAMD": "College of Arts, Media and Design",
                "CE": "College of Engineering",
                "CPS": "College of Professional Studies",
                "CS": "College of Science",
                "CSSH": "College of Social Sciences and Humanities",
                "DMSB": "D'Amore-McKim School of Business",
                "KCCS": "Khoury College of Computer Sciences",
                "SL": "School of Law"
            },
            "x-enum-varnames": [
                "CAMD",
                "DMSB",
                "KCCS",
                "CE",
                "BCHS",
                "SL",
                "CPS",
                "CS",
                "CSSH"
            ]
        },
...
        "models.RecruitmentCycle": {
            "type": "string",
            "enum": [
                "fall",
                "spring",
                "fallSpring",
                "always"
            ],
            "x-enum-varnames": [
                "Fall",
                "Spring",
                "FallSpring",
                "Always"
            ]
        },
        "models.RecruitmentType": {
            "type": "string",
            "enum": [
                "unrestricted",
                "tryout",
                "application"
            ],
            "x-enum-varnames": [
                "Unrestricted",
                "Tryout",
                "Application"
            ]
        },
...
        "models.Tag": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "college",
                "email",
                "first_name",
                "last_name",
                "nuid",
                "role",
                "year"
            ],
            "properties": {
                "clubs_followed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Club"
                    }
                },
                "college": {
                    "maxLength": 255,
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.College"
                        }
                    ]
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "nuid": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "super",
                        "student"
                    ]
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "year": {
                    "maximum": 6,
                    "minimum": 1,
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Year"
                        }
                    ]
                }
            }
        },
        "models.Year": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6
            ],
            "x-enum-varnames": [
                "First",
                "Second",
                "Third",
                "Fourth",
                "Fifth",
                "Graduate"
            ]
        },
...
    }
}